+++ compiler-compat/k230/src/main/kotlin/dev/zacsweers/metro/compiler/compat/k230/CompatContextImpl.kt
@@ -99,6 +99,10 @@ import org.jetbrains.kotlin.util.PrivateForInline
 
 public class CompatContextImpl : CompatContext {
 
+  override fun createCompilerConfigurationCompat(): CompilerConfiguration {
+    return CompilerConfiguration()
+  }
+
   override fun KtSourcelessDiagnosticFactory.createCompat(
     message: String,
     location: CompilerMessageSourceLocation?,