+++ compiler-compat/k240/src/main/kotlin/dev/zacsweers/metro/compiler/compat/k240/CompatContextImpl.kt
@@ -10,7 +10,9 @@ import kotlin.reflect.KClass
 import org.jetbrains.kotlin.backend.common.extensions.DeclarationFinder
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
+import org.jetbrains.kotlin.cli.create
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
+import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.diagnostics.KtDiagnosticFactory1
 import org.jetbrains.kotlin.fir.FirAnnotationContainer
 import org.jetbrains.kotlin.fir.FirElement
@@ -53,6 +55,10 @@ import org.jetbrains.kotlin.name.Name
 
 public class CompatContextImpl : CompatContext by DelegateType() {
   override val supportsAnnotationArgumentInvalidation: Boolean = true
+
+  override fun createCompilerConfigurationCompat(): CompilerConfiguration {
+    return CompilerConfiguration.create()
+  }
 
   context(_: CompilerPluginRegistrar)
   override fun CompilerPluginRegistrar.ExtensionStorage.registerFirExtensionCompat(